src/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
async-property==0.2.2
    # via python-keycloak
asyncpg==0.30.0
    # via -r requirements.in
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via click
cryptography==44.0.0
    # via jwcrypto
deprecation==2.1.0
    # via python-keycloak
ecdsa==0.19.0
    # via python-jose
fastapi==0.115.6
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via python-keycloak
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jwcrypto==1.5.6
    # via python-keycloak
packaging==24.2
    # via deprecation
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via pydantic-settings
python-jose==3.3.0
    # via -r requirements.in
python-keycloak==5.1.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
rsa==4.9
    # via python-jose
ruff==0.8.4
    # via -r requirements.in
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via -r requirements.in
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   jwcrypto
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r requirements.in
